perf(loader-gltf): lazy-load KHR_animation_pointer material registry
Node-only pointer assets (e.g. CubeVisibility / scene34, which only animates
KHR_node_visibility) were bundling the entire material pointer registry plus the
KHR_materials_* ext-texture resolver (anisotropy/sheen/specular/transmission)
they never use. Split material (\/materials/...\) pointer resolution into a new
gltf-pointer-material feature, loaded by the registry only when a channel targets
a material; node TRS + visibility now resolve inline in the eager pointer feature.
gltf-animation consults the lazily-installed material resolver after the node
parser returns null.

scene34 94.6 -> 90.5 KB raw (back under its 91 KB ceiling). Material-pointer
scenes pay a small chunk-split overhead. All 12 animation-pointer parity scenes
unchanged (scene34 MAD 0.000, scene241 0.188).
